added test Stub for RawDiagonal
